chore: 파일 변경사항

+++ src/main/java/com/sprint/hrbank/application/employee/provided/query/EmployeeCounter.java
@@ -2,5 +2,5 @@ package com.sprint.hrbank.application.employee.provided.query;
 
 public interface EmployeeCounter {
 
-  Integer countEmployeesByDepartment_Id(Long departmentId);
+  Long countEmployeesByDepartment_Id(Long departmentId);
 }
